diffusionlight-github/cloud-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
supabase==2.3.0
postgrest==0.13.2
storage3==0.7.4
gotrue==2.3.0
realtime==1.0.4

